Add accumulate to tree
This algorithm aims to replace the current compile-time AccumulateValue and AccumulateType meta-programs. In particular, it aims to be more readable, more flexible, and usable at run-time.
This in turn, also allows to change the tree info algorithm to a very simple visitor. Base on those, we test that the algorithm works as expected.