Add support for non-trivial multi-index literals
The existing multi-index (and former tree-path) literal is a little
point-less, since numeric literals only alllow for multi-indices with
a single digit. As of C++20 one can have string literal operator
templates. These allow to pass (with a tiny hack) string literals
as template parameters. Hence we can utilize this to define non-trivial
multi-index literals, too. E.g.

```cpp
auto a = 10_mi; // Same as hybridMultiIndex(_10);
auto a = ""_mi; // Same as hybridMultiIndex();
auto a = "10"_mi; // Same as hybridMultiIndex(_10);
auto a = "0, 1, 2"_mi; // Same as hybridMultiIndex(_0, _1, _2);
```

The hack required by the standard is to capture a string literal
in a class instance which can be used as template parameter. The
corresponding helper class is a general utility and might be suited
for dune-common, since it can be used to implement other string
literal operator tempates, too.